+++ packages/map-template/src/components/Map/Map.jsx
@@ -20,6 +20,7 @@ import pitchState from '../../atoms/pitchState';
 import solutionState from '../../atoms/solutionState';
 import notificationMessageState from '../../atoms/notificationMessageState';
 import useMapBoundsDeterminer from '../../hooks/useMapBoundsDeterminer';
+import currentVenueNameState from "../../atoms/currentVenueNameState";
 
 /**
  * Private variable used for storing the tile style.
@@ -53,6 +54,7 @@ function Map({ onLocationClick, onVenueChangedOnMap, useMapProviderModule, onMap
     const [, setPositionControl] = useRecoilState(positionControlState);
     const solution = useRecoilValue(solutionState);
     const [, setErrorMessage] = useRecoilState(notificationMessageState);
+    const [, setCurrentVenueName] = useRecoilState(currentVenueNameState);
     useLiveData(apiKey);
 
     const [mapPositionKnown, venueOnMap] = useMapBoundsDeterminer();
@@ -127,9 +129,11 @@ function Map({ onLocationClick, onVenueChangedOnMap, useMapProviderModule, onMap
 
     /*
      * When venue is changed on the map, run callback.
+     * Set the current venue name whenever the venue is changed on the map.
      */
     useEffect(() => {
         if (venueOnMap) {
+            setCurrentVenueName(venueOnMap.name);
             onVenueChangedOnMap(venueOnMap);
         }
     }, [venueOnMap]);