+++ packages/map-template/CHANGELOG.md
@@ -5,6 +5,31 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.43.1] - 2024-04-08
+
+### Fixed
+
+- Fix the venue not updating correctly on the map. 
+
+## [1.43.0] - 2024-03-26
+
+### Added
+
+- Add support for new parameter `searchAllVenues` which searches across all venues in one solution. 
+- Change the default behaviour to search in one venue.
+
+## [1.42.0] - 2024-03-26
+
+### Added
+
+- Enhance the visual representation of `non-selectable` locations in the search results list.
+
+## [1.41.0] - 2024-03-19
+
+### Added
+
+- Added Custom Element definition to Web Component generation file. This means you can avoid inline scripting, and only load a single JS-file from a CDN to get up and running with the Map Template Web Component
+
 ## [1.40.0] - 2024-03-06
 
 ### Added